requirements.txt
streamlit
pillow
matplotlib
numpy
opencv-python